+++ cmp-worker-kmp/src/jvmMain/kotlin/io/github/mobilebytelabs/worker/PlatformContext.jvm.kt
@@ -1,0 +1,3 @@
+package io.github.mobilebytelabs.worker
+
+public actual class PlatformContext public constructor()